+++ Sources/StreamVideo/protobuf/ProtoModel.swift
@@ -20,3 +20,4 @@ extension Stream_Video_Sfu_Signal_ICETrickleResponse: ErrorProviding {}
 extension Stream_Video_Sfu_Signal_UpdateSubscriptionsResponse: ErrorProviding {}
 extension Stream_Video_Sfu_Signal_UpdateMuteStatesResponse: ErrorProviding {}
 extension Stream_Video_Sfu_Signal_ICERestartResponse: ErrorProviding {}
+extension Stream_Video_Sfu_Signal_SendStatsResponse: ErrorProviding {}